==> 01_Notebooks/01_Data_Collection/01_Scrape_API_Event_IDs.ipynb <==
# Scrape API Event IDs

## Imports

import pandas as pd
import requests
import json
import re
import time
from bs4 import BeautifulSoup
import os
import errno

!ls ../../../../Data/events.csv

events = pd.read_csv('../../../../Data/events.csv', index_col=0)

urls = events.url
base_url = 'http://www.ufc.com'

list_of_event_dicts = []

# First create header for the csv
with open('event_ids.csv', 'a+', newline='') as f:
    f.write(f'url,event_id')

for num, url in enumerate(urls):

    time.sleep(0.5)
    full_url = base_url + url
    response = requests.get(full_url)
    response_text = response.text
    try:
        event_val = int(re.findall('/V1/(\d{3})/Fnt\.js', response_text)[0])
    except:
        event_val = 0
    event_dict = {}
    event_dict['event_id'] = event_val
    event_dict['event_url'] = url
    list_of_event_dicts.append(event_dict)
    with open('event_ids.csv', 'a+', newline='') as f:
        f.write(f'\n{url}, {event_val}')
