==> 01_Notebooks/01_Data_Collection/02_Scrape_API_Fight_IDs_and_V1_Jsons.ipynb <==
# Scrape API Fight IDs

## Imports

import pandas as pd
import requests
import json
import re
import time
from bs4 import BeautifulSoup
import os
import errno
import time
import datetime

def scrape_all_event_fight_ids(events_df):
    create_fight_stat_ids_csv()
    for index, event in events_df.iterrows():
#         if index%5==0:
#             print(f'{index}: {event.url}')
        
        url = event.url
        event_id = event.event_id
        list_of_fight_stat_ids = scrape_fight_stat_ids_from_one_event(event_id)
        try:
            if len(list_of_fight_stat_ids) > 0:
                write_fight_stat_ids_to_csv(url, event_id, list_of_fight_stat_ids)
        except:
            print(f"There was a write error for {event.url}")
        time.sleep(3)

def scrape_fight_stat_ids_from_one_event(event_id):
    full_url = f'http://liveapi.fightmetric.com/V1/{event_id}/Fnt.json'
    
    try:
        response = requests.get(full_url)
        structured = json.loads(response.text)
        list_of_fight_stat_ids = [x['FightID'] for x in structured['FMLiveFeed']['Fights']]
    except:
        list_of_fight_stat_ids = []
    return list_of_fight_stat_ids

def create_fight_stat_ids_csv():
    # Create the header
    try:
        os.remove('event_fight_stat_ids.csv')
    except:
        pass
    
    with open('event_fight_stat_ids.csv', 'a+', newline='') as f:
        f.write(f'url,event_id,fight_stat_id')

def write_fight_stat_ids_to_csv(url, event_id, list_of_fight_stat_ids):        
    for fight_stat_id in list_of_fight_stat_ids:
        with open('event_fight_stat_ids.csv', 'a+', newline='') as f:
            f.write(f'\n{url},{event_id},{fight_stat_id}')

## Pull in Event IDs and Scrape

!ls ../Scraped_Pre_Data

events = pd.read_csv('../Scraped_Pre_Data/event_ids.csv')
events = events[events.event_id != 0]
events.head()

scrape_all_event_fight_ids(events)

# Update this to also pull the full JSON for each V1 url

def scrape_all_v1_urls(events_df):
    for index, event in events_df.iterrows():   
        event_url = event.url
        event_id = event.event_id    
        if index%5==0:
            print(f'{index} of {len(events_df)}: {event_id}') 
        if event_id != 0:
            scrape_one_v1_urls(event_id)
        else: 
            print(f"skipping {index}-{event_url}")
        time.sleep(1)

def scrape_one_v1_urls(event_id):
    v1_url = f'http://liveapi.fightmetric.com/V1/{event_id}/Fnt.json'
    
    response = requests.get(v1_url)
    
    if response.status_code == 200:    
        data = response.json()
        
        file_path = f"../Scraped_V1_Jsons/v1_event_{event_id}.json"
        directory = os.path.dirname(file_path)

        with open(file_path, 'w') as f:
            json.dump(data, f)
    else:
        log_v1_response_error(event_id)

def log_v1_response_error(event_id):
    print("there was a response error")
    with open('error_log_v1_scrape.csv', 'a+', newline='') as f:
        f.write(f'\n**********************************************')
        f.write(f'\nProblem Type:  Response Error')
        f.write(f'\nDateTime:  {datetime.datetime.now().strftime("%Y-%m-%d %H:%M")}')
        f.write(f'\nEvent_id: {event_id}')

!ls ../Scraped_V1_Jsons

scrape_one_v1_urls(866)

scrape_all_v1_urls(events)
