==> 01_Notebooks/01_Data_Collection/03_Scrape_API_V2_Fight_Jsons.ipynb <==
# 03 Scrape API Fight Jsons

## Imports

import pandas as pd
import requests
import json
import re
import time
from bs4 import BeautifulSoup
import os
import errno
import time
import datetime

def scrape_all_fight_jsons(fight_stat_ids_df):
    for key, row in fight_stat_ids_df.iterrows():
        time.sleep(1)

        event_id = row.event_id
        fight_stat_id = row.fight_stat_id
        v2_url = f'http://liveapi.fightmetric.com/V2/{event_id}/{fight_stat_id}/Stats.json'
        
        if key%50==0:
            print(f'scraping: {key} of {len(fight_stat_ids_df)}')
        
        response = requests.get(v2_url)

        if response.status_code == 200:
            data = response.json()

            file_path = f"./Scraped_Data/event_{event_id}/fight_{fight_stat_id}.json"
            directory = os.path.dirname(file_path)

            try:
                os.makedirs(directory)
            except OSError as e:
                if e.errno != errno.EEXIST:
                    raise

            file_path = f"./Scraped_Data/event_{event_id}/fight_{fight_stat_id}.json"
            with open(file_path, 'w') as f:
                json.dump(data, f)
        else:
            log_v2_response_error(event_id, fight_stat_id)

def log_v2_response_error(event_id, fight_stat_id):
    print("there was a response error")
    with open('error_log.csv', 'a+', newline='') as f:
        f.write(f'\n**********************************************')
        f.write(f'\nProblem Type:  Response Error')
        f.write(f'\nDateTime:  {datetime.datetime.now().strftime("%Y-%m-%d %H:%M")}')
        f.write(f'\nEvent_id: {event_id}')
        f.write(f'\nfight_stat_id: {fight_stat_id}')   

## Pull in Fight_IDs and Scrape Fight Jsons

fight_stat_ids_df = pd.read_csv('../Scraped_Pre_Data/event_fight_stat_ids.csv')

scrape_all_fight_jsons(fight_stat_ids_df)
