==> 01_Notebooks/01_Data_Collection/04_Scrape_UFC_Fighter_Page_Info.ipynb <==
# 01 Scraping Fighter Info

## Imports

import pandas as pd
import requests
from bs4 import BeautifulSoup
import math
import re
from matplotlib import pyplot as plt
import time
import datetime

## Get List of Urls for Each Page of Fighters

url = 'http://www.ufc.com/fighter/Weight_Class/filterFighters?offset={offset}&fighterFilter=All'
response = requests.get(url)
soup = BeautifulSoup(response.content, 'lxml')
number_of_fighters = int(soup.find('div', {'class':'paginate-results'}).find_all('span', {'class':'row-count'})[1].text)
fighters_per_page = 20
pages = math.ceil(number_of_fighters/fighters_per_page)

urls = []
for i in range(pages):
    offset = i*20
    url = f'http://www.ufc.com/fighter/Weight_Class/filterFighters?offset={offset}&fighterFilter=All'
    urls.append(url)

list_of_fighter_dict = []

for page_num, url in enumerate(urls):

    response = requests.get(url)
    soup = BeautifulSoup(response.content, 'lxml')
    
    # Get the list of fighters object from the page
    fighters_html = soup.find('table', {'class':'fighter-listing'})
    list_of_fighter_htmls = fighters_html.find_all('tr', {'class':'fighter'})
    
    if page_num%5==0:
        print(f"Scrapping Page {page_num } of {len(urls)}...")
    for fighter_html in list_of_fighter_htmls:
        fighter_dict = {}
        fighter_dict['fighter_slug'] = fighter_html.find('div', {'class':'fighter-info'}).find('a').attrs['href'][9:]
        fighter_dict['fighter_name'] = re.sub('\n', '',fighter_html.find('div', {'class':'fighter-info'}).find('a').text.strip())
        
        if len(fighter_html.find_all('div', {'class':'main-txt'})) == 3:
            # Get wins, losses and ties
            win_loss_html = fighter_html.find_all('div', {'class':'main-txt'})[0]
            if win_loss_html.text:
                fighter_dict['wins'] = int(re.match("(\d*)-(\d*)-(\d*)", win_loss_html.text).group(1))
                fighter_dict['losses'] = int(re.match("(\d*)-(\d*)-(\d*)", win_loss_html.text).group(2))
                fighter_dict['ties'] = int(re.match("(\d*)-(\d*)-(\d*)", win_loss_html.text).group(3))

            # Get Height - problem - sometimes no height

            height_html = fighter_html.find_all('div', {'class':'main-txt'})[1]

            feet = int(re.match("(\d*)' (\d*)", height_html.text).group(1))
            inches = int(re.match("(\d*)' (\d*)", height_html.text).group(2))
            total_height = feet*12 + inches
            fighter_dict['height'] = total_height

            # Get Weight
            weight_html = fighter_html.find_all('div', {'class':'main-txt'})[2]
            fighter_dict['weight'] = int(re.match("(\d*) lbs", weight_html.text).group(1))

        list_of_fighter_dict.append(fighter_dict)
        time.sleep(1)
    print("Done!")

df = pd.DataFrame(list_of_fighter_dict)

len(list_of_fighter_dict)

## Export Fighter Slug Data

df.to_csv('../../02_Data/01_Raw_Scraped_Data/Fighters_Slug.csv')

## For each fighter in the list, go to their UFC page and pull down their info

def get_all_fighter_info(list_fighter_slugs):
    list_fighter_info_dict = []
    
    for key, fighter_slug in enumerate(list_fighter_slugs):
        if key%25==0:
            print(f'{key}: Scraping {fighter_slug} at {datetime.datetime.now().strftime("%Y-%m-%d %H:%M")}')
        fighter_url = f'http://www.ufc.com/fighter/{fighter_slug}'
        try:
            response = requests.get(fighter_url)
            soup = BeautifulSoup(response.content, 'lxml')
            fighter_info_dict = get_fighter_info(soup, fighter_slug)
            list_fighter_info_dict.append(fighter_info_dict)

        except:
            log_scraping_fights(fighter_slug)
            
        time.sleep(1)
        
    return list_fighter_info_dict

def log_scraping_fights(fighter_slug):
    print(f'error with {fighter_slug} at {datetime.datetime.now().strftime("%Y-%m-%d %H:%M")}')
    with open('fighter_info_error_log.csv', 'a+', newline='') as f:
        f.write(f'\nerror scraping {fighter_slug} at {datetime.datetime.now().strftime("%Y-%m-%d %H:%M")}')


def get_fighter_info(soup, fighter_slug):
    fighter_info_dict = {}
    for tr in soup.find('div', {'class':'fighter-info'}).find('table').find_all('tr'):
        label = tr.find('td', {'class':'label'}).text
        value = tr.find('td', {'class':'value'}).text

        # Try to clean up the value if exists
        try:
            value = re.sub('\n','', value)
            value = re.sub('\t','', value)
        except:
            print('something here??')
            pass
        fighter_info_dict[label] = value
    fighter_info_dict['fighter_slug'] = fighter_slug
    return fighter_info_dict

list_fighter_info_dict = get_all_fighter_info(df.fighter_slug)

df = pd.DataFrame(list_fighter_info_dict)

## Export

df.to_csv('../../02_Data/01_Raw_Scraped_Data/Fighter_Info2181.csv')
